Cleanup and finish merge
