+++ src/core/webview/webviewMessageHandler.ts
@@ -3881,25 +3881,7 @@ export const webviewMessageHandler = async (
 		// kilocode_change end
 		// kilocode_change start - ManagedIndexer state
 		case "requestManagedIndexerState": {
-			try {
-				ManagedIndexer.getInstance()?.sendStateToWebview()
-			} catch (error) {
-				provider.log(
-					`Error getting managed indexer state: ${error instanceof Error ? error.message : String(error)}`,
-				)
-				await provider.postMessageToWebview({
-					type: "managedIndexerState",
-					managedIndexerEnabled: false,
-					managedIndexerState: [],
-				})
-			}
-			break
-		}
-
-		// we're going to delete this message at some point and use apiConfiguration
-		// probably. So casting as any as to not define a more permanent type
-		case "requestManagedIndexerEnabled" as any: {
-			ManagedIndexer.getInstance()?.sendEnabledStateToWebview()
+			ManagedIndexer.getInstance()?.sendStateToWebview()
 			break
 		}
 		// kilocode_change end
