+++ webview-ui/src/components/chat/kilocode/TabbedCodeIndexPopover.tsx
@@ -1,24 +1,23 @@
 // kilocode_change - new file
-import React, { useState, useRef } from "react"
-import { Trans } from "react-i18next"
+import React, { useState, useRef, useEffect } from "react"
 import { VSCodeLink } from "@vscode/webview-ui-toolkit/react"
 
 import type { IndexingStatus } from "@roo/ExtensionMessage"
 
-import { useAppTranslation } from "@src/i18n/TranslationContext"
-import { buildDocLink } from "@src/utils/docLinks"
-import { Popover, PopoverContent, Tabs, TabsContent, TabsList, TabsTrigger } from "@src/components/ui"
+import { Popover, PopoverContent } from "@src/components/ui"
 import { useRooPortal } from "@src/components/ui/hooks/useRooPortal"
 import { useEscapeKey } from "@src/hooks/useEscapeKey"
-import { CodeIndexPopover } from "../CodeIndexPopover"
-import { ManagedCodeIndexPopoverContent } from "./ManagedCodeIndexPopover"
+import { vscode } from "@src/utils/vscode"
+import { ManagedIndexerStatus } from "./ManagedIndexerStatus"
+import { type WorkspaceFolderState } from "./managedIndexerSchema"
+import { useExtensionState } from "../../../context/ExtensionStateContext"
 
 interface CodeIndexPopoverProps {
 	children: React.ReactNode
 	indexingStatus: IndexingStatus
 }
 
-export const TabbedCodeIndexPopover: React.FC<CodeIndexPopoverProps> = ({ children, indexingStatus }) => {
+export const TabbedCodeIndexPopover: React.FC<CodeIndexPopoverProps> = ({ children }) => {
 	const [open, setOpen] = useState(false)
 	const closeHandlerRef = useRef<() => void>(() => setOpen(false))
 
@@ -50,53 +49,51 @@ export const TabbedCodeIndexPopover: React.FC<CodeIndexPopoverProps> = ({ childr
 				collisionPadding={16}
 				avoidCollisions={true}
 				container={portalContainer}>
-				<TabbedCodeIndexPopoverTabs
-					indexingStatus={indexingStatus}
-					setOpen={setOpen}
-					onRegisterCloseHandler={(handler) => {
-						// We use this pattern so the CodeIndexPopover can decide if it wants
-						// the popover to close based on if it has unsaved changes
-						closeHandlerRef.current = handler
-					}}
-				/>
+				<TabbedCodeIndexPopoverTabs />
 			</PopoverContent>
 		</Popover>
 	)
 }
 
-export const TabbedCodeIndexPopoverTabs = ({
-	indexingStatus,
-	setOpen,
-	onRegisterCloseHandler,
-}: {
-	indexingStatus: IndexingStatus
-	setOpen: (open: boolean) => void
-	onRegisterCloseHandler: (handler: () => void) => void
-}) => {
-	const { t } = useAppTranslation()
-	const [activeTab, setActiveTab] = useState("managed")
+const TabbedCodeIndexPopoverTabs = () => {
+	const state = useExtensionState()
+	const orgId = state.apiConfiguration?.kilocodeOrganizationId
+	const href = `https://kilo.ai/organizations/${orgId}/code-indexing`
+	const [workspaceFolders, setWorkspaceFolders] = useState<WorkspaceFolderState[]>([])
+
+	// Request initial state when popover opens
+	// Listen for managed indexer state updates
+	useEffect(() => {
+		console.log("[ManagedCodeIndexPopoverContent] requesting managed indexer state")
+		vscode.postMessage({ type: "requestManagedIndexerState" })
+
+		const handleMessage = (event: MessageEvent<any>) => {
+			console.log("[ManagedCodeIndexPopoverContent] received event", event)
+			if (event.data.type === "managedIndexerState") {
+				setWorkspaceFolders(event.data.managedIndexerState || [])
+			}
+		}
+
+		window.addEventListener("message", handleMessage)
+		return () => window.removeEventListener("message", handleMessage)
+	}, [])
 
 	return (
-		<Tabs value={activeTab} onValueChange={setActiveTab}>
+		<>
 			<div className="p-3 cursor-default">
 				<div className="flex flex-row items-center gap-1 p-0 mt-0 mb-1 w-full">
-					<h4 className="m-0 pb-2 flex-1">{t("settings:codeIndex.title")}</h4>
+					<h4 className="m-0 pb-2 flex-1">Managed Code Indexing</h4>
 				</div>
 				<p className="my-0 pr-4 text-sm w-full mb-3">
-					<Trans i18nKey="settings:codeIndex.description">
-						<VSCodeLink
-							href={buildDocLink("features/codebase-indexing", "settings")}
-							style={{ display: "inline" }}
-						/>
-					</Trans>
+					<VSCodeLink href={href}>Configure on kilo.ai</VSCodeLink>
 				</p>
 			</div>
 
 			<div className="border-t border-vscode-dropdown-border" />
 
 			<div className="p-4">
-				<ManagedCodeIndexPopoverContent />
+				<ManagedIndexerStatus workspaceFolders={workspaceFolders} />
 			</div>
-		</Tabs>
+		</>
 	)
 }
